// src/models/public/PlaceReportValues.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import type { AccountsAccountId } from './Accounts';
import type { PlaceReportsPlaceReportId } from './PlaceReports';
import type { UnitsUnitId } from './Units';

/** Identifier type for public.place_report_values */
export type PlaceReportValuesPlaceReportValueId = string & { __brand: 'public.place_report_values' };

/**
 * Represents the table public.place_report_values
 * value-ing the situation of the subproject in this place
 */
export default interface PlaceReportValues {
  place_report_value_id: PlaceReportValuesPlaceReportValueId;

  /** redundant account_id enhances data safety */
  account_id: AccountsAccountId | null;

  place_report_id: PlaceReportsPlaceReportId | null;

  unit_id: UnitsUnitId | null;

  /** Used for integer values */
  value_integer: number | null;

  /** Used for numeric values */
  value_numeric: number | null;

  /** Used for text values */
  value_text: string | null;

  label: string | null;

  created_at: Date;

  updated_at: Date;

  updated_by: string | null;
}

/**
 * Represents the initializer for the table public.place_report_values
 * value-ing the situation of the subproject in this place
 */
export interface PlaceReportValuesInitializer {
  /** Default value: uuid_generate_v7() */
  place_report_value_id?: PlaceReportValuesPlaceReportValueId;

  /** redundant account_id enhances data safety */
  account_id?: AccountsAccountId | null;

  place_report_id?: PlaceReportsPlaceReportId | null;

  unit_id?: UnitsUnitId | null;

  /** Used for integer values */
  value_integer?: number | null;

  /** Used for numeric values */
  value_numeric?: number | null;

  /** Used for text values */
  value_text?: string | null;

  label?: string | null;

  /** Default value: now() */
  created_at?: Date;

  /** Default value: now() */
  updated_at?: Date;

  updated_by?: string | null;
}

/**
 * Represents the mutator for the table public.place_report_values
 * value-ing the situation of the subproject in this place
 */
export interface PlaceReportValuesMutator {
  place_report_value_id?: PlaceReportValuesPlaceReportValueId;

  /** redundant account_id enhances data safety */
  account_id?: AccountsAccountId | null;

  place_report_id?: PlaceReportsPlaceReportId | null;

  unit_id?: UnitsUnitId | null;

  /** Used for integer values */
  value_integer?: number | null;

  /** Used for numeric values */
  value_numeric?: number | null;

  /** Used for text values */
  value_text?: string | null;

  label?: string | null;

  created_at?: Date;

  updated_at?: Date;

  updated_by?: string | null;
}